Create STATA Codes used for Contingent Valuation Method Analysis
There are two section of the codes. The current section is the first part of the code. Here in the first part the univariate analysis carried out by preparation of charts and correlation matrix. The goal of uploading these codes to make the viewer understand CVM analysis from a household survey.